fixed file delete

+++ routes/files.js
@@ -29,20 +29,31 @@ router.post('/:file', function(req, res) {
 router.delete('/:file', function(req, res) {
 	if (req.user) {
 		User.findById(req.user._id, (err, user) => {
-     let ownfileid= user.ownFiles[user.ownFiles.findIndex((element, index, array) => element.nameHash === req.params.file)]
-          .nameHash
-          let secondfileid =user.secondFiles[user.secondFiles.findIndex((element, index, array) => element.nameHash === req.params.file)]
-          .nameHash;
-			websockets.deleteFile(ownfileid?ownfileid:secondfileid)
+			let ownfile =
+				user.ownFiles[user.ownFiles.findIndex((element, index, array) => element.nameHash === req.params.file)]
+			let ownfileid = ownfile && ownfile.nameHash
+			let secondfile =
+				user.secondFiles[
+					user.secondFiles.findIndex((element, index, array) => element.nameHash === req.params.file)
+				]
+			let secondfileid = secondfile && secondfile.nameHash
+			if (!ownfileid && !secondfileid) return
 
-      
+			websockets.deleteFile(ownfileid ? ownfileid : secondfileid)
 
-
-			user.ownFiles.splice(
-				user.ownFiles.findIndex((element, index, array) => element.nameHash === req.params.file),
-				1
-			)
+			if (ownfileid) {
+				user.ownFiles.splice(
+					user.ownFiles.findIndex((element, index, array) => element.nameHash === req.params.file),
+					1
+				)
+			} else {
+				user.secondFiles.splice(
+					user.secondFiles.findIndex((element, index, array) => element.nameHash === req.params.file),
+					1
+				)
+			}
 			user.save()
+			res.sendStatus(200)
 		})
 	} else {
 		res.sendStatus(401)
